Extract charge classifier into its own class
